photo-clustering-app/requirements.txt
streamlit>=1.30.0
insightface>=0.7.3
onnxruntime>=1.16.0
umap-learn>=0.5.5
hdbscan>=0.8.33
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
gdown>=5.1.0
